Refactor navbar code
